fix: extract Redis and Hazelcast auto-configurations to top-level classes
Spring Boot ignores @AutoConfiguration ordering annotations on nested
classes. The RedisConfiguration inner class had
@AutoConfiguration(after = DataRedisAutoConfiguration.class) but as a
nested class this ordering was not applied, causing RedisConnector to
never be created because RedisConnectionFactory did not exist yet when
the condition was evaluated.

Extract RedisConfiguration and HazelcastConfiguration to top-level
auto-configuration classes registered in AutoConfiguration.imports so
their ordering is respected. Extract HazelcastProperties to its own
@ConfigurationProperties class with the same prefix for consistency.
Add @AutoConfigureAfter for both on KubernetesKitConfiguration so the
backend connectors exist when @ConditionalOnBean(BackendConnector) is
evaluated.
