+++ kubernetes-kit-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,3 @@
 com.vaadin.kubernetes.starter.KubernetesKitConfiguration
+com.vaadin.kubernetes.starter.RedisConfiguration
+com.vaadin.kubernetes.starter.HazelcastConfiguration
